make common test utils dir; restructure util_libs
