+++ rust/clients/orchestrator/Cargo.toml
@@ -28,6 +28,7 @@ utoipa-swagger-ui = { version = "9", features = [
   "vendored",
 ] }
 utoipa = { version = "5", features = ["actix_extras"] }
-util_libs = { path = "../../util_libs" }
+nats_utils = { path = "../../util_libs/nats" }
+db_utils = { path = "../../util_libs/db" }
 workload = { path = "../../services/workload" }
 inventory = { path = "../../services/inventory" }
