+++ rust/services/inventory/src/types.rs
@@ -1,7 +1,7 @@
 use hpos_hal::inventory::HoloInventory;
+use nats_utils::types::{CreateResponse, CreateTag, EndpointTraits};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use util_libs::nats::types::{CreateResponse, CreateTag, EndpointTraits};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum InventoryPayloadType {
