+++ rust/services/workload/Cargo.toml
@@ -26,12 +26,13 @@ url = { version = "2", features = ["serde"] }
 bytes = "1.8.0"
 nkeys = "=0.4.4"
 chrono = "0.4.0"
-util_libs = { path = "../../util_libs" }
+db_utils = { path = "../../util_libs/db" }
+nats_utils = { path = "../../util_libs/nats" }
 hpos-hal = { path = "../../hpos-hal" }
 
 [dev-dependencies]
-tempfile = "3.8"
 serde_json = "1.0"
+mock_utils = { path = "../../util_libs/mocks" }
 
 [features]
 tests_integration_workload_service = []
