+++ rust/services/workload/src/types.rs
@@ -1,11 +1,9 @@
 use bson::oid::ObjectId;
+use db_utils::schemas::{self, WorkloadStatus};
+use nats_utils::types::{CreateResponse, CreateTag, EndpointTraits};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use strum_macros::AsRefStr;
-use util_libs::{
-    db::schemas::{self, WorkloadStatus},
-    nats::types::{CreateResponse, CreateTag, EndpointTraits},
-};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct ObjectIdJSON {
