+++ rust/util_libs/db/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "util_libs"
+name = "db_utils"
 version = "0.1.0"
 edition = "2021"
 
@@ -17,21 +17,19 @@ log = { workspace = true }
 env_logger = { workspace = true }
 thiserror = { workspace = true }
 dotenv = { workspace = true }
+mongodb = "3.1.0"
+chrono = "0.4.0"
+bytes = "1.8.0"
+tempfile = "3.8"
 tokio-retry = "0.3.0"
 async-trait = "0.1.83"
+strum = "0.24"
+sea-strum_macros = "0.23.0"
 url = { version = "2", features = ["serde"] }
-mongodb = "3.1.0"
 bson = { version = "2.6.1", features = ["chrono-0_4"] }
 derive_more = { version = "1.0.0", features = ["debug"] }
-chrono = "0.4.0"
-sea-strum_macros = "0.23.0"
-strum = "0.24"
-bytes = "1.8.0"
-hpos-hal = { path = "../hpos-hal" }
-
-[dev-dependencies]
-tempfile = "3.14"
+hpos-hal = { path = "../../hpos-hal" }
+nats_utils = { path = "../../util_libs/nats" }
 
 [features]
 tests_integration_mongodb = []
-tests_integration_nats = []
