+++ rust/util_libs/nats/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "nats_utils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+async-nats = { workspace = true }
+anyhow = { workspace = true }
+nats-jwt = "0.3.0"
+serde = { workspace = true }
+serde_json = { workspace = true }
+serde_with = { version = "3.1", features = ["macros"] }
+semver = "1.0.24"
+futures = { workspace = true }
+tokio = { workspace = true }
+log = { workspace = true }
+env_logger = { workspace = true }
+thiserror = { workspace = true }
+dotenv = { workspace = true }
+mongodb = "3.1.0"
+chrono = "0.4.0"
+bytes = "1.8.0"
+tempfile = "3.8"
+tokio-retry = "0.3.0"
+async-trait = "0.1.83"
+strum = "0.24"
+sea-strum_macros = "0.23.0"
+url = { version = "2", features = ["serde"] }
+bson = { version = "2.6.1", features = ["chrono-0_4"] }
+derive_more = { version = "1.0.0", features = ["debug"] }
+hpos-hal = { path = "../../hpos-hal" }
+
+[features]
+tests_integration_nats = []
