--- rust/util_libs/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod db;
-pub mod nats;
